Merge FH changes July/August (second try)

See merge request Zurich_ttH/tthbb13!22

+++ MEAnalysis/python/nanoTree.py
@@ -7,7 +7,7 @@ class EventAnalyzer(Analyzer):
         super(EventAnalyzer, self).__init__(cfg_ana, cfg_comp, looperName)
     def process(self, event):
         event.lumi = getattr(event.input, "luminosityBlock", None)
-    	event.evt = getattr(event.input, "event", None)
+    	event.evt = long(getattr(event.input, "event", None))
         #event.json = getattr(event.input, "json", None)
         #event.json_silver = getattr(event.input, "json_silver", None)
